Missing test conversion to Cascading 3.0
