+++ parquet-cascading3/src/test/java/org/apache/parquet/cascading3/TestParquetTBaseScheme.java
@@ -1,4 +1,4 @@
-/* 
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.parquet.cascading;
+package org.apache.parquet.cascading3;
 
 import cascading.flow.Flow;
 import cascading.flow.FlowProcess;
@@ -33,11 +33,13 @@ import cascading.tap.hadoop.Hfs;
 import cascading.tuple.Fields;
 import cascading.tuple.Tuple;
 import cascading.tuple.TupleEntry;
+
 import org.apache.commons.io.FileUtils;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.BytesWritable;
+import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
 import org.apache.hadoop.mapreduce.TaskAttemptID;
 import org.apache.thrift.protocol.TCompactProtocol;
@@ -45,78 +47,82 @@ import org.apache.thrift.protocol.TProtocol;
 import org.apache.thrift.protocol.TProtocolFactory;
 import org.apache.thrift.transport.TIOStreamTransport;
 import org.junit.Test;
+import static org.junit.Assert.*;
+
 import org.apache.parquet.hadoop.thrift.ThriftToParquetFileWriter;
 import org.apache.parquet.hadoop.util.ContextUtil;
 import org.apache.parquet.thrift.test.Name;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.ByteArrayOutputStream;
+import java.util.HashMap;
+import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
-
-public class TestParquetTupleScheme {
-  final String parquetInputPath = "target/test/ParquetTupleIn/names-parquet-in";
-  final String txtOutputPath = "target/test/ParquetTupleOut/names-txt-out";
-
-  @Test
-  public void testReadPattern() throws Exception {
-    String sourceFolder = parquetInputPath;
-    testReadWrite(sourceFolder);
-
-    String sourceGlobPattern = parquetInputPath + "/*";
-    testReadWrite(sourceGlobPattern);
-
-    String multiLevelGlobPattern = "target/test/ParquetTupleIn/**/*";
-    testReadWrite(multiLevelGlobPattern);
-  }
+public class TestParquetTBaseScheme {
+  final String txtInputPath = "src/test/resources/names.txt";
+  final String parquetInputPath = "target/test/ParquetTBaseScheme/names-parquet-in";
+  final String parquetOutputPath = "target/test/ParquetTBaseScheme/names-parquet-out";
+  final String txtOutputPath = "target/test/ParquetTBaseScheme/names-txt-out";
 
   @Test
-  public void testFieldProjection() throws Exception {
-    createFileForRead();
-
-    Path path = new Path(txtOutputPath);
-    final FileSystem fs = path.getFileSystem(new Configuration());
+  public void testWrite() throws Exception {
+    Path path = new Path(parquetOutputPath);
+    JobConf jobConf = new JobConf();
+    final FileSystem fs = path.getFileSystem(jobConf);
     if (fs.exists(path)) fs.delete(path, true);
 
-    Scheme sourceScheme = new ParquetTupleScheme(new Fields("last_name"));
-    Tap source = new Hfs(sourceScheme, parquetInputPath);
+    Scheme sourceScheme = new TextLine( new Fields( "first", "last" ) );
+    Tap source = new Hfs(sourceScheme, txtInputPath);
 
-    Scheme sinkScheme = new TextLine(new Fields("last_name"));
-    Tap sink = new Hfs(sinkScheme, txtOutputPath);
+    Scheme sinkScheme = new ParquetTBaseScheme(Name.class);
+    Tap sink = new Hfs(sinkScheme, parquetOutputPath);
 
-    Pipe assembly = new Pipe("namecp");
-    assembly = new Each(assembly, new ProjectedTupleFunction());
-    Flow flow = new HadoopFlowConnector().connect("namecp", source, sink, assembly);
+    Pipe assembly = new Pipe( "namecp" );
+    assembly = new Each(assembly, new PackThriftFunction());
+    HadoopFlowConnector hadoopFlowConnector = new HadoopFlowConnector();
+    Flow flow  = hadoopFlowConnector.connect("namecp", source, sink, assembly);
 
     flow.complete();
-    String result = FileUtils.readFileToString(new File(txtOutputPath + "/part-00000"));
-    assertEquals("Practice\nHope\nHorse\n", result);
+
+    assertTrue(fs.exists(new Path(parquetOutputPath)));
+    assertTrue(fs.exists(new Path(parquetOutputPath + "/_metadata")));
+    assertTrue(fs.exists(new Path(parquetOutputPath + "/_common_metadata")));
   }
 
-  public void testReadWrite(String inputPath) throws Exception {
+  @Test
+  public void testRead() throws Exception {
+    doRead(new ParquetTBaseScheme(Name.class));
+  }
+
+  @Test
+  public void testReadWithoutClass() throws Exception {
+    doRead(new ParquetTBaseScheme());
+  }
+
+  private void doRead(Scheme sourceScheme) throws Exception {
     createFileForRead();
 
     Path path = new Path(txtOutputPath);
     final FileSystem fs = path.getFileSystem(new Configuration());
     if (fs.exists(path)) fs.delete(path, true);
 
-    Scheme sourceScheme = new ParquetTupleScheme(new Fields("first_name", "last_name"));
-    Tap source = new Hfs(sourceScheme, inputPath);
+    Tap source = new Hfs(sourceScheme, parquetInputPath);
 
     Scheme sinkScheme = new TextLine(new Fields("first", "last"));
     Tap sink = new Hfs(sinkScheme, txtOutputPath);
 
-    Pipe assembly = new Pipe("namecp");
-    assembly = new Each(assembly, new UnpackTupleFunction());
-    Flow flow = new HadoopFlowConnector().connect("namecp", source, sink, assembly);
+    Pipe assembly = new Pipe( "namecp" );
+    assembly = new Each(assembly, new UnpackThriftFunction());
+    Flow flow  = new HadoopFlowConnector().connect("namecp", source, sink, assembly);
 
     flow.complete();
-    String result = FileUtils.readFileToString(new File(txtOutputPath + "/part-00000"));
+    String result = FileUtils.readFileToString(new File(txtOutputPath+"/part-00000"));
     assertEquals("Alice\tPractice\nBob\tHope\nCharlie\tHorse\n", result);
   }
 
+
   private void createFileForRead() throws Exception {
-    final Path fileToCreate = new Path(parquetInputPath + "/names.parquet");
+    final Path fileToCreate = new Path(parquetInputPath+"/names.parquet");
 
     final Configuration conf = new Configuration();
     final FileSystem fs = fileToCreate.getFileSystem(conf);
@@ -150,32 +156,30 @@ public class TestParquetTupleScheme {
     w.close();
   }
 
-  private static class UnpackTupleFunction extends BaseOperation implements Function {
+  private static class PackThriftFunction extends BaseOperation implements Function {
     @Override
     public void operate(FlowProcess flowProcess, FunctionCall functionCall) {
       TupleEntry arguments = functionCall.getArguments();
       Tuple result = new Tuple();
 
-      Tuple name = new Tuple();
-      name.addString(arguments.getString(0));
-      name.addString(arguments.getString(1));
+      Name name = new Name();
+      name.setFirst_name(arguments.getString(0));
+      name.setLast_name(arguments.getString(1));
 
       result.add(name);
       functionCall.getOutputCollector().add(result);
     }
   }
 
-  private static class ProjectedTupleFunction extends BaseOperation implements Function {
+  private static class UnpackThriftFunction extends BaseOperation implements Function {
     @Override
     public void operate(FlowProcess flowProcess, FunctionCall functionCall) {
       TupleEntry arguments = functionCall.getArguments();
       Tuple result = new Tuple();
 
-      Tuple name = new Tuple();
-      name.addString(arguments.getString(0));
-//      name.addString(arguments.getString(1));
-
-      result.add(name);
+      Name name = (Name) arguments.get(0);
+      result.add(name.getFirst_name());
+      result.add(name.getLast_name());
       functionCall.getOutputCollector().add(result);
     }
   }
